+++ src/main/java/Models/Position.java
@@ -1,5 +1,5 @@
 package Models;
 
 public interface Position {
-    int getPosition();
+    long getPosition();
 }
